requirements.txt
allure_python_commons==2.8.0
DingtalkChatbot==1.5.2
pytest==6.2.4
requests==2.24.0
allpairspy==2.5.0
allure-pytest==2.8.0
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==9.1.1
pytest-xdist==1.30.0